requirements.txt
alembic==1.12.0
blinker==1.6.3
click==8.1.7
colorama==0.4.6
dnspython==2.4.2
email-validator==2.0.0.post2
Flask==2.2.5
Flask-Cors==4.0.0
Flask-Login==0.6.2
flask-marshmallow==0.15.0
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
greenlet==3.0.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.0
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.3
marshmallow==3.20.2
marshmallow-SQLAlchemy==0.26.1
packaging==23.2
psycopg2-binary==2.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
six==1.15.0
SQLAlchemy==2.0.27
typing_extensions==4.9.0
Werkzeug==2.3.7
WTForms==3.0.1
zipp==3.12.0